Crate cumulus_primitives_aura

Core primitives for Aura in Cumulus.

In particular, this exposes the AuraUnincludedSegmentApi which is used to regulate the behavior of Aura within a parachain context.

Structs

Slot

Unit type wrapper that represents a slot.

Traits

AuraUnincludedSegmentApi

This runtime API is used to inform potential block authors whether they will have the right to author at a slot, assuming they have claimed the slot.